+++ front-end/package.json
@@ -5,7 +5,7 @@
   "private": true,
   "homepage": "http://polkassembly.io/",
   "dependencies": {
-    "@polkadot/api": "^1.9.0-beta.7",
+    "@polkadot/api": "^1.8.1",
     "@polkadot/extension-dapp": "^0.22.1",
     "@polkadot/extension-inject": "^0.22.1",
     "@polkadot/react-identicon": "^0.52.0-beta.6",
